refactor: :coffin: rm Status column, it's only confusing
Status is almost always '(active)' which might imply running to some people, when in fact it only means that it's the world in 'worldstorage' and not 'worldstorage_inactive', which is a legacy carryover from the bash scripts implementation we might remove. instead, it's easier to think of 1:1 servers to worlds and just create more servers as needed.